+++ parser/parser_predicates.c
@@ -5,7 +5,7 @@ bool is_normalized(t_vec3 vec)
 	return (float_eq(vec3_length(vec), 1));
 }
 
-bool in_interval(float value, float min, float max)
+bool in_interval(double value, double min, double max)
 {
 	return (value >= min && value <= max);
 }
